Refactor import statements in SignatureVerifier.sol and ECDSA.sol

+++ src/SignatureVerifier.sol
@@ -5,9 +5,6 @@ pragma solidity ^0.8.24;
 import "./libary/ERC20.sol";
 import "./libary/ECDSA.sol";
 
-
-
-
 contract SignatureVerifier {
     using ECDSA for bytes32;
 
@@ -17,19 +14,18 @@ contract SignatureVerifier {
 
     constructor(address _token, address[] memory _whitelist) {
         token = IERC20(_token);
-        for (uint i = 0; i < _whitelist.length; i++) {
+        for (uint256 i = 0; i < _whitelist.length; i++) {
             whitelist[_whitelist[i]] = true;
         }
     }
 
- function verifyAndClaim(bytes32 messageHash, bytes memory signature) external {
+    function verifyAndClaim(bytes32 messageHash, bytes memory signature) external {
         address signer = messageHash.recover(signature);
         require(whitelist[signer], "Address not whitelisted");
         require(signer == msg.sender, "Signer must be the sender");
         require(!hasClaimed[signer], "Tokens already claimed");
 
         hasClaimed[signer] = true;
-        require(token.transfer(signer, 100 * 10**18), "Token transfer failed");
+        require(token.transfer(signer, 100 * 10 ** 18), "Token transfer failed");
     }
-   
 }
